% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kntn_unnest.R
\name{kntn_unnest_records}
\alias{kntn_unnest_records}
\title{Unnest 'kintone' Records}
\usage{
kntn_unnest_records(records)
}
\arguments{
\item{records}{Data.frame retrived by kntn_records()}
}
\description{
The data retrieved by \link{kntn_records} may contain nested data.frames.
\code{kntn_unnest_records()} unnests them by using \link[tidyr]{unnest}.
Note that this function is very experimental and may not work well for all data.
}

